% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chloromethane}
\alias{chloromethane}
\title{Chloromethane calibration data using GC-MS}
\format{
A data frame with 90 rows and 2 variables:
\describe{
\item{Concentration}{μg/L}
\item{Response}{Arbitrary unit}
}
}
\source{
{I. Lavagnini and F. Magno, (2007), Mass Spec. Rev., 26:1-18}
}
\usage{
data(chloromethane)
}
\description{
Chloromethane (CH3Cl) calibration data using GC-MS with fluorobenzene as an internal standard (IS)
}
\details{
A dataset containing chloromethane concentration (μg/L, x) and the ratio of the measured peak
area for chloromethane to that of the internal standard, flurobenzene.
}
\keyword{datasets}
