% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hplccal}
\alias{hplccal}
\title{HPLC calibration data for five analytes.}
\format{
A data frame with 70 rows and three variables:
\describe{
\item{Analyte}{}
\item{Concentration}{Different units}
\item{Area}{Arbitrary unit}
}
}
\source{
{L. Kirkup and M. Mulholland, (2004), J. Chrom. A, 1029:1-11}
}
\usage{
data(hplccal)
}
\description{
Calibration data for five analytes (ibuprofen, genisten, biochanin,
pseudoephedrine and sodium nitrate) using different instrumental
conditions.
}
\keyword{datasets}
