% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paraquat}
\alias{paraquat}
\title{Paraquat calibration data using multiple square wave voltammetry.}
\format{
A data frame with 9 rows and 2 variables:
\describe{
\item{x}{μmol/L}
\item{y}{μA (X -1)}
}
}
\source{
{O.B. da Silva and S.A.S. Machado, (2012), Anal. Methods, 4:2348-2354}
}
\usage{
data(paraquat)
}
\description{
A dataset containing concentration (μmol/L, x) and measured response (μA X -1, y).
}
\keyword{datasets}
