% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPlot.R
\name{plot.lod}
\alias{plot.lod}
\title{Plot method for 'lod' objects}
\usage{
\method{plot}{lod}(x, residplot = NULL, ...)
}
\arguments{
\item{x}{a 'lod' object}

\item{residplot}{flag for regression residuals plot}

\item{...}{optional parameters}
}
\description{
Plot function to view calibration curve, based on linear,
quadratic or power regression.
}
\examples{
data(mtbe)
m <- det.limit(mtbe$x, mtbe$y) # linear regression
plot(m)

}
