% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPrint.R
\name{print.lod}
\alias{print.lod}
\title{Print output from 'lod' model}
\usage{
\method{print}{lod}(x, dp = NULL, ...)
}
\arguments{
\item{x}{a 'lod' object}

\item{dp}{number of decimal points}

\item{...}{optional parameters}
}
\description{
Print output from 'lod' model
}
\examples{
data(mtbe)

m <- det.limit(mtbe$x, mtbe$y) # linear regression

print(m, dp = 2) # two decimal points
}
