% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vhtab6}
\alias{vhtab6}
\title{Calibration data from Table 6, Vogelsang et al (1998).}
\format{
A data frame with 12 rows and 2 variables:
\describe{
\item{concentration}{μg/kg}
\item{area}{arbitrary units}
}
}
\source{
{J. Vogelgesang and J. Hädrich (1998), Accred. Qual. Assur., 3:242-255.}
}
\usage{
data(vhtab6)
}
\description{
A sample dataset of concentration (μg/kg) and measured response (area).
}
\keyword{datasets}
